vm/spider/requirements.txt
python-dotenv
azure-storage-blob==2.1.0